refactor(services): move store-inject-config
