+++ src/pages/todo-list/page/children/create-list-item/index.tsx
@@ -2,14 +2,9 @@ import React from 'react';
 import { RouteNode } from '@wildberries/service-router';
 import { ReduxStoreLoader } from '@mihanizm56/redux-core-modules';
 import { AppLayout } from '@/_layouts/app-layout';
-import {
-  CREATE_ITEM_WATCHER_SAGA_NAME,
-  TODO_LIST_REDUCER_NAME,
-} from '@/pages/todo-list/_redux/todo-list';
-import { createItemWatcherSaga } from '@/pages/todo-list/_redux/todo-list/sagas/create-item-watcher-saga';
-import reducerTodoList from '@/pages/todo-list/_redux/todo-list/reducer';
 import { CreatePage } from './page';
 import { CREATE_ITEM_PAGE_PAGE_NODE } from './_constants';
+import { storeInjectConfig } from './store-inject-config';
 
 const pageNode = CREATE_ITEM_PAGE_PAGE_NODE;
 
@@ -23,20 +18,7 @@ const action = async ({ fromState, toState }) => ({
             return (
               <ReduxStoreLoader
                 fromState={fromState}
-                storeInjectConfig={{
-                  reducersToInject: [
-                    {
-                      name: TODO_LIST_REDUCER_NAME,
-                      reducer: reducerTodoList,
-                    },
-                  ],
-                  sagasToInject: [
-                    {
-                      name: CREATE_ITEM_WATCHER_SAGA_NAME,
-                      saga: createItemWatcherSaga,
-                    },
-                  ],
-                }}
+                storeInjectConfig={storeInjectConfig()}
                 toState={toState}
               >
                 <CreatePage />
